Add Rules to NSG
This script creates inbound allow rules for common Windows services in an Azure Network Security Group (NSG). It prompts the user to select a resource group and an NSG within that group, then allows the user to select which services to create rules for. The rules are created with incrementing priorities starting at 110.